backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pip-autoremove==0.10.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
setuptools==76.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
Pillow